be/AI/requirements.txt
faster-whisper>=0.8.0
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
